// src/generated/models/cinema.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * Cinema Backend API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Кинотеатр
 */
export interface Cinema {
  /** Идентификатор кинотеатра */
  id?: number;
  /** Название кинотеатра */
  name?: string;
  /** Адрес кинотеатра */
  address?: string;
}