// src/generated/models/movieSession.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * Cinema Backend API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Киносеанс
 */
export interface MovieSession {
  /** Идентификатор сеанса */
  id?: number;
  /** Идентификатор фильма */
  movieId?: number;
  /** Идентификатор кинотеатра */
  cinemaId?: number;
  /** Время начала сеанса */
  startTime?: string;
}